[REFACTOR] 음주 목표 실제 음주기록 반영 #32

+++ src/main/java/com/umc/puppymode2/domain/goal/dto/GoalInfoResponseDTO.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @Builder
 public class GoalInfoResponseDTO {
     private Integer monthlyGoalCount;
-    private Integer monthlyActualCount;
+    private Long monthlyActualCount;
     private Boolean isGoalExceeded;
     private LocalDateTime goalSetAt;
 }
